+++ app/backend/src/mareas/mareas.service.ts
@@ -1883,6 +1883,12 @@ export class MareasService {
                                 }
                             }
                         }
+                    },
+                    {
+                        metadata: {
+                            path: ['observadoresSecundariosPlanificados'],
+                            array_contains: [{ observadorId: observadorId }]
+                        }
                     }
                 ]
             },